network/requirements.txt
keras==2.3.1
scikit-learn==1.2.2
numpy==1.18.5
joblib==1.2.0
nltk==3.8.1

pandas==1.4.0

Flask==2.3.1
Flask_RESTful==0.3.9

Pillow==9.5.0
pymorphy2==0.9.1
python-decouple==3.8
regex==2023.3.23
requests==2.29.0
beautifulsoup4==4.12.2
